refactor(tests): B1 tier relocation — PG-requiring suites leave the unit tier + ci-data-v5
Moves (git mv, history preserved):
- test_per_tick_transaction_atomicity, test_trace_view_columns{,_v2},
  test_migration_apply_healing: tests/unit/persistence ->
  tests/integration/persistence (each REQUIRES live Postgres; the unit
  tier is import-light/mock-based by ADR074 rule 1)
- tests/unit/persistence/conftest.py -> tests/integration/persistence/
  migration_healing.py (pure helpers, no fixtures — pg_pool/pg_dsn come
  from the root conftest); importers updated.

The move immediately caught a FOURTH raw migration loop: the atomicity
file's own apply_062_migrations raced the idempotency suite on the
pg_class catalog (UniqueViolation ix_hex_state_session_tick) the moment
they became xdist neighbors — replaced with ensure_ddl_applied. The
whole integration persistence tier: 20 passed in 52s; combined with the
remaining (mock-based) unit tier: 422 passed.

Also:
- test_county_aggregation: 3 reference-DB classes marked
  requires_reference_db (refdata lane routing; skipif guards intact)
- fetch-reference-db composite bumped to ci-data-v5 (released:
  fact_fred_wealth_shares FULL + fact_ricci_unequal_exchange populated;
  sha256 2e7920e2...) — keeps the refdata CI legs green with the new
  refdb sync tests
- two battery-verified triage fixes that predate the compaction ride
  here: conservation_audit_strict patched_advance **extra forwarding;
  hex_spatial_map_isolation PK-order query via unnest WITH ORDINALITY
  (the old ORDER BY k.n referenced a nonexistent pg_index column)
