+++ tests/integration/persistence/test_trace_view_columns.py
@@ -55,7 +55,7 @@ def view_applied(pg_pool):  # type: ignore[no-untyped-def]
     mid-migration by a killed background pytest run is healed loudly with one
     retry instead of erroring fixture setup — see conftest.py docstring.
     """
-    from tests.unit.persistence.conftest import apply_migrations_healing
+    from tests.integration.persistence.migration_healing import apply_migrations_healing
 
     apply_migrations_healing(pg_pool)
